util: Extract flush_icache_range to cacheflush.c
This has been a tcg-specific function, but is also in use
by hardware accelerators via physmem.c.  This can cause
link errors when tcg is disabled.

Message-Id: <20201214140314.18544-3-richard.henderson@linaro.org>
(cherry picked from commit 084cfca143487d9b3ef37e7ee117f30e8e301af1)
